Rename .H to .HPP also for WinMFC
