+++ VisualStudio/AirspaceConverterMFC/LimitsDlg.hpp
